v1.0.4 - updated dependency to fermiummixins
rlmixins had the overwrites for broken heart and undershirt before, now its fermiummixins

todo: fermiummixins is not on curseforge yet, gotta update the version (gradle.properties) to curse maven once its on cf.
