+++ build.gradle
@@ -58,7 +58,7 @@ dependencies {
     implementation fg.deobf("curse.maven:fermiumbooter-${fermiumbooter_version}")
     implementation fg.deobf("curse.maven:reskillable-${reskillable_version}")
     implementation fg.deobf("curse.maven:bountifulbaubles-${bountifulbaubles_version}")
-    implementation fg.deobf("curse.maven:rlmixins-${rlmixins_version}")
+    implementation fg.deobf("libs:FermiumMixins:${fermiummixins_version}")
     implementation fg.deobf("curse.maven:firstaid-${firstaid_version}")
 }
