+++ gradle.properties
@@ -22,7 +22,7 @@ mixin_extras_version=0.3.6
 # mod
 mod_id=bhutracker
 mod_name=Broken Heart & Undershirt Stat
-mod_version=1.0.3
+mod_version=1.0.4
 mod_author=["Iqury","Nischhelm"]
 mod_description=Adds a Broken Heart and Undershirt activated count
 mixin_loader_class=bhutracker.MixinPlugin
@@ -31,5 +31,6 @@ mixin_loader_class=bhutracker.MixinPlugin
 fermiumbooter_version=968474:5672830
 reskillable_version=286382:2815686
 bountifulbaubles_version=313536:3568240
-rlmixins_version=711730:6046850
+# todo: update to cursemaven once fermiummixins is on cf
+fermiummixins_version=1.0.0
 firstaid_version=276837:4030769
