+++ src/main/java/bhutracker/BrokenHeartUndershirtTracker.java
@@ -17,7 +17,7 @@ import net.minecraftforge.fml.common.event.FMLPreInitializationEvent;
 )
 public class BrokenHeartUndershirtTracker {
     public static final String MODID = "bhutracker";
-    public static final String VERSION = "1.0.3";
+    public static final String VERSION = "1.0.4";
     public static final String NAME = "Broken Heart & Undershirt Stat";
 
     public static StatBase UNDERSHIRT_ACTIVATED = (new StatBasic("stat.undershirt_activated", new TextComponentTranslation(TextFormatting.AQUA + "Undershirt Activated"))).registerStat();
