+++ src/main/resources/mixins.bhutracker.fermiummixins.json
@@ -1,6 +1,6 @@
 {
   "required": true,
-  "package": "bhutracker.mixin.rlmixins",
+  "package": "bhutracker.mixin.fermiummixins",
   "refmap": "mixins.bhutracker.refmap.json",
   "compatibilityLevel": "JAVA_8",
   "target": "@env(DEFAULT)",
